Move menu code into separate file and folder
  * Now in menu/mainmenu.js.
